fix(build): move compat sidecars into src
